+++ src/Commands/DriveDistanceCommand.h
@@ -39,34 +39,34 @@
  */
 class DriveDistanceCommand: public Command {
 private:
-	float distanceL;
-	float distanceR;
+	double distanceL;
+	double distanceR;
 	bool isCommandDone;
-	float distanceError;
-	float remainingDistance;
-	float voltageLeft = 0.0;
-	float voltageRight = 0.0;
-	float maxVoltage = 0.8;
-	float timeStep = 0.02;
-	float secondsToMax = 1.5;
+	double distanceError;
+	double remainingDistance;
+	double voltageLeft = 0.0;
+	double voltageRight = 0.0;
+	double maxVoltage = 0.8;
+	double timeStep = 0.02;
+	double secondsToMax = 1.5;
 	// Just make voltageStep 1.0 and let the voltage ramp rate handle it.
-	//float voltageStep = maxVoltage/(secondsToMax/timeStep);
-	float voltageStep = 1.0;
-	float wheelDiam = 6.25;
-	float distancePerRev = 3.1416*wheelDiam;
-	float inchesPerTick = distancePerRev/7680;
+	//double voltageStep = maxVoltage/(secondsToMax/timeStep);
+	double voltageStep = 1.0;
+	double wheelDiam = 6.25;
+	double distancePerRev = 3.1416*wheelDiam;
+	double inchesPerTick = distancePerRev/7680;
 	// If we're off by 1 inch, increase the voltage by 0.05
-	float voltageScale = 0.05;
+	double voltageScale = 0.05;
 	int ticks = 0;
 public:
-	DriveDistanceCommand(float distL, float distR);
+	DriveDistanceCommand(double distL, double distR);
 	virtual void Initialize();
 	virtual void Execute();
 	virtual bool IsFinished();
 	virtual void End();
 	virtual void Interrupted();
-	float distanceTravelledL;
-	float distanceTravelledR;
+	double distanceTravelledL;
+	double distanceTravelledR;
 };
 
 #endif
